+++ src/main/java/com/skala/decase/domain/project/mapper/ProjectMapper.java
@@ -1,6 +1,5 @@
 package com.skala.decase.domain.project.mapper;
 
-import com.skala.decase.domain.document.domain.Document;
 import com.skala.decase.domain.member.domain.Member;
 import com.skala.decase.domain.project.controller.dto.request.CreateProjectRequest;
 import com.skala.decase.domain.project.controller.dto.response.MappingTableResponseDto;
@@ -10,7 +9,6 @@ import com.skala.decase.domain.project.domain.Project;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-import com.skala.decase.domain.requirement.controller.dto.RequirementDto;
 import com.skala.decase.domain.requirement.domain.Requirement;
 import com.skala.decase.domain.requirement.domain.RequirementDocument;
 import lombok.AllArgsConstructor;
