+++ src/main/java/com/skala/decase/domain/requirement/service/RequirementService.java
@@ -5,9 +5,9 @@ import com.skala.decase.domain.member.exception.MemberException;
 import com.skala.decase.domain.member.repository.MemberRepository;
 import com.skala.decase.domain.project.domain.Project;
 import com.skala.decase.domain.project.service.ProjectService;
-import com.skala.decase.domain.requirement.controller.dto.RequirementDto;
-import com.skala.decase.domain.requirement.controller.dto.RequirementRevisionDto;
-import com.skala.decase.domain.requirement.controller.dto.UpdateRequirementDto;
+import com.skala.decase.domain.requirement.controller.dto.request.RequirementDto;
+import com.skala.decase.domain.requirement.controller.dto.request.RequirementRevisionDto;
+import com.skala.decase.domain.requirement.controller.dto.request.UpdateRequirementDto;
 import com.skala.decase.domain.requirement.controller.dto.response.RequirementWithSourceResponse;
 import com.skala.decase.domain.requirement.domain.Requirement;
 import com.skala.decase.domain.requirement.exception.RequirementException;
